# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: HCML Monitoring Portal (BFF)
  version: 1.0.0
  summary: A read-only backend-for-frontend over the Zabbix 7.0 JSON-RPC API.
  description: |
    The portal owns **no database**. Every response below is either passed through from Zabbix or
    derived in-process and cached in a 500-entry LRU that is lost on restart. Sites, links,
    incidents, host reachability and the monthly SLA have no table anywhere.

    **One write exists** in the whole surface: `POST /api/problems/acknowledge`. It uses a separate
    write-capable Zabbix token and is refused with 503 when that token is unset.

    ### Authentication

    Two schemes, and both work on every guarded path:

    - `Authorization: Bearer <jwt>`, the normal one, from `POST /api/auth/login`. 12-hour expiry.
    - `?token=<jwt>`: a query-string alternative. It exists because `EventSource` cannot set a
      header, but `auth.ts:129-132` reads it on **every** `/api/*` request, not only the SSE ones.
      `logging.ts` redacts it from the request log.

    When `AUTH_ENABLED=false` no route is guarded at all and every caller is treated as `admin`.

    ### Roles

    OpenAPI cannot express a role requirement per path, so each operation states its role in the
    description. From `ROUTE_RULES` (`server/src/auth.ts:35-40`), first match wins, unmatched is
    `viewer`:

    | Role | Paths |
    |---|---|
    | `admin` | `/api/reports/inventory` |
...
    | `viewer` | everything else |

    ### Errors

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/net/devices:
    get:
      summary: Network devices with ICMP state
      description: '**operator.** Cached 30 s. Narrowed to `NET_GROUP_IDS` when that is configured.'
      operationId: getNetDevices
      tags: [Network]
      responses:
        '200':
          description: Devices.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/NetDevice' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/net/ports:
    get:
      summary: Port items for one device
      description: '**operator.** Cached 15 s. About 1,100 rows on HCML''s core switch.'
      operationId: getNetPorts
      tags: [Network]
      parameters:
        - name: hostid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: Raw `item.get` rows.
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/net/interfaces:
    get:
      summary: Interfaces for one device, searched then filtered then paged
      description: |
        **operator.** Cached 15 s.

        Order matters and is observable: `search` is applied first, `summary` is computed over the
        searched set, *then* `status` filters, *then* the page is sliced. So `summary` counts more
        rows than `rows` contains whenever `status` is set.

        **`status` fails silently.** Any value outside the enum is treated as no filter rather than
        rejected with 400.
      operationId: getNetInterfaces
      tags: [Network]
      parameters:
        - name: hostid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100000, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 10, maximum: 1000, default: 200 }
        - name: search
          in: query
          description: Case-insensitive substring of `name` or `alias`.
          schema: { type: string }
        - name: status
          in: query
          description: An unrecognised value is ignored, not rejected.
          schema: { type: string, enum: [up, down, other] }
      responses:
        '200':
          description: One page of interfaces.
          content:
            application/json:
              schema:
                type: object
                required: [rows, total, summary, page, pageSize]
                properties:
                  rows: { type: array, items: { $ref: '#/components/schemas/InterfaceRow' } }
                  total: { type: integer }
                  summary:
                    type: object
                    required: [up, down, other]
                    properties:
                      up: { type: integer }
                      down: { type: integer }
                      other: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/net/status:
    get:
      summary: Interface status items for one device
      description: '**operator.** Cached 15 s. Not called by the web app today.'
      operationId: getNetStatus
      tags: [Network]
      parameters:
        - name: hostid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: Raw `item.get` rows.
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/net/map:
    get:
      summary: One Zabbix map, for the network view
      description: '**operator. Not cached.** Not called by the web app today.'
      operationId: getNetMap
      tags: [Network]
      parameters:
        - name: mapid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: Raw `map.get` result.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The normal scheme. 12-hour expiry, carries `{ sub, role }`.
    queryToken:
      type: apiKey
      in: query
      name: token
      description: |
        The same JWT in the query string. Required for `EventSource`, which cannot set headers, but
        accepted on **every** guarded path. Kept out of the request log by `logging.ts`.
  responses:
    BadRequest:
      description: '`bad_request`, a parameter failed validation.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: '`unauthorized`, missing or expired portal JWT. Not the same as `zabbix_auth`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: '`forbidden`, authenticated, but the role is below what this path requires.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ZabbixError:
      description: '`zabbix_error`, unreachable, non-2xx, non-JSON-RPC, or a JSON-RPC error.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ZabbixUnavailable:
      description: |
        `zabbix_timeout` (no answer within `ZABBIX_TIMEOUT_MS`, default 10 s) or `zabbix_auth` (Zabbix
        rejected the API token). Both are configuration states rather than upstream faults, which is
        why they are 503 and not 504.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error, message]
      description: The only error shape this server produces, at every status.
      properties:
        error: { type: string, description: A stable machine-readable code. }
        message: { type: string, description: Safe to show a user. Never contains the Zabbix URL. }
...
    SiteRef:
      type: object
      description: A resolved site. No Zabbix entity corresponds to this.
      properties:
        code: { type: string }
        name: { type: string }
      additionalProperties: true
...
    NetDevice:
      type: object
      required: [hostid, name, status]
      properties:
        hostid: { type: string }
        name: { type: string }
        status: { type: string }
        site:
          oneOf: [{ $ref: '#/components/schemas/SiteRef' }, { type: 'null' }]
        icmp:
          oneOf:
            - type: object
              properties:
                up: { type: boolean }
                loss: { type: number }
                latency: { type: number }
                state: { type: string }
            - type: 'null'
    InterfaceRow:
      type: object
      required: [name, oper]
      properties:
        name: { type: string }
        alias: { type: string }
        oper: { type: string, enum: [up, down, other] }
        speed: { type: number }
        inOctets: { type: number }
        outOctets: { type: number }
      additionalProperties: true
...